requirements.txt
altair~=5.4.1
beautifulsoup4~=4.12.3
bs4~=0.0.2
gnews~=0.3.7
google-api-python-client~=2.149.0
google-play-scraper~=1.2.7
joblib~=1.3.2
langdetect~=1.0.9
lxml_html_clean
matplotlib~=3.8.2
newspaper3k~=0.2.8
nltk~=3.8.1
numpy~=1.26.3
pandas~=2.1.4
pillow~=10.2.0
python-dotenv~=1.0.1
requests~=2.31.0
seaborn~=0.13.1
setuptools~=61.0.0
spotipy==2.24.0
squarify~=0.4.4
streamlit~=1.38.0
Unidecode~=1.3.8
youtube-scraping-api~=1.1.4
pytest